requirements.txt
learn2learn>=0.1.7
lightning-bolts>=0.5.0
numpy>=1.19.5
Pillow>=8.4.0
tensorboard>=2.8.0
timm @ git+https://github.com/rwightman/pytorch-image-models@a520da9b495422bc773fb5dfe10819acb8bd7c5c
torch>=1.11.0+cu113
torchaudio>=0.11.0+cu113
torchvision>=0.12.0+cu113
tqdm>=4.62.3
pandas>=1.3.4